Prevent retrieving the same feed in parallel (#366)
* Prevent retrieving the same feed in parallel

* typo

+++ includes/class-feed.php
@@ -113,7 +113,18 @@ class Feed {
 	 * Cron function to refresh the feeds of the friends' blogs
 	 */
 	public function cron_friends_refresh_feeds() {
-		$this->retrieve_friend_posts();
+		foreach ( User_Feed::get_all_due() as $feed ) {
+			if ( ! $feed->can_be_polled_now() ) {
+				continue;
+			}
+			$feed->set_polling_now();
+			$this->retrieve_feed( $feed );
+			$feed->was_polled();
+			$friend_user = $feed->get_friend_user();
+			if ( $friend_user ) {
+				$friend_user->delete_outdated_posts();
+			}
+		}
 	}
 
 	/**
@@ -129,7 +140,8 @@ class Feed {
 
 		foreach ( $friend_user->get_active_feeds() as $feed ) {
 			$friend_user = $feed->get_friend_user();
-			if ( $friend_user ) {
+			if ( $friend_user && $feed->can_be_polled_now() ) {
+				$feed->set_polling_now();
 				$this->retrieve_feed( $feed );
 				$feed->was_polled();
 				$friend_user->delete_outdated_posts();
@@ -280,10 +292,11 @@ class Feed {
 	/**
 	 * Retrieve posts from all friends.
 	 *
-	 * @param      bool $also_undue     Whether to get also undue feeds.
+	 * @param      bool $ignore_due_date     Whether to get also undue feeds.
 	 */
-	public function retrieve_friend_posts( $also_undue = false ) {
-		foreach ( User_Feed::get_all_due( $also_undue ) as $feed ) {
+	public function retrieve_friend_posts( $ignore_due_date = false ) {
+		foreach ( User_Feed::get_all_due( $ignore_due_date ) as $feed ) {
+			$feed->set_polling_now();
 			$this->retrieve_feed( $feed );
 			$feed->was_polled();
 			$friend_user = $feed->get_friend_user();
